Fixed some warning regarding the variable types for x64. CinemaCell+Cinema.m was actually a category on uitableviewcell, fixed that, and also CinemaCell wasn't being used in storyboard. Changed font for the label in cinemacell. Fixed warning about info.plit being included in copy bundle resources.

+++ Cine Horarios/Classes/AppDelegate.m
@@ -106,7 +106,7 @@ static int const kGaDispatchPeriod = 30;
     NSDictionary *userInfo = [notification userInfo];
     NSString *theaterName = [userInfo valueForKey:@"TheaterName"];
     NSUInteger theaterID = [[userInfo valueForKey:@"TheaterID"] integerValue];
-    NSString *key = [NSString stringWithFormat:@"%d",theaterID];
+    NSString *key = [NSString stringWithFormat:@"%lu",(unsigned long)theaterID];
     
     NSArray *paths = NSSearchPathForDirectoriesInDomains (NSDocumentDirectory, NSUserDomainMask, YES);
     NSString *documentsPath = [paths objectAtIndex:0];
